Minor refactors to simplify review of following commits

+++ core/builtins/src/kotlin/Progressions.kt
@@ -96,7 +96,7 @@ public open class IntProgression
     /**
      * The last element in the progression.
      */
-    public val last: Int = getProgressionLastElement(start.toInt(), endInclusive.toInt(), step).toInt()
+    public val last: Int = getProgressionLastElement(start, endInclusive, step).toInt()
 
     /**
      * The step of the progression.
@@ -150,7 +150,7 @@ public open class LongProgression
     /**
      * The last element in the progression.
      */
-    public val last: Long = getProgressionLastElement(start.toLong(), endInclusive.toLong(), step).toLong()
+    public val last: Long = getProgressionLastElement(start, endInclusive, step).toLong()
 
     /**
      * The step of the progression.
